Slightly different hack
Dave about the previous hack: While it does appear to work, it's doing something that it not ideal. If you open the network panel, you will see GET https://wbazant.plasmodb.org/plasmo.wbazant/service/users/current/datasets/NaN. This is because you are setting the value of the dataset param to PF3D7_0321200 , but a dataset param value should be a dataset id. Internally, the dataset param is converting the value to a number object. Number(PF3D7_0321200) returns NaN, which is why you see that request. The upshot is that we should not use the idList as the param value.The reason it appears to work is that the datasetParam module is still using the initialParamData object to initialize its state. So it sees params.ds_gene_ids.idList=PF3D7_0321200 and populates the textarea with that value.I think an alternative is to just use "" as the value of a dataset param for the purpose of POSTing to the search endpoint.