Fix bug about order of actions: should first post the file, then transition
Wojtek: I want the page to wait until the file body is posted, but then transition. But actually, that's not what it does - it transitions sometime between the first request (start upload, sets status as "awaiting upload") and the second. So, it leads to a screen which says "awaiting upload" while still posting the file (which I see for a larger file, when I open a network tab) which is not exactly a bug, but not how I wanted it either - I wanted the "new upload" screen to be at low opacity until the file is posted, and transition afterwards.
The bug scenario (found today :slightly_smiling_face: ) is that when the user refreshes the page (and it's very tempting, there's a button to refresh the page, etc.) the upload gets stuck as "awaiting upload" forever: the file does not complete loading, and the server never fails it.

Dave: you are not returning the result of fetchWithCredentials
you can fix that with an explicit return, or by removing the braces, which creates an implicit return