Prevent render when ownProps and data are out-of-sync.
This can happen when going from one record type to another, and is
especially problematic when there is an intermediate non-record-page. In
this case, the initial render will use state from the previous page,
which will cause partial record requests to be made for the previous
record type. Since ownProps and data props are out of sync, requests can
contain fields incompatible with the record type.