Display star icon for multifilter variables in visualization variable picker.
Part of this change involves modifying how we indicate that multifilter descendant variables should be displayed in the variable tree. Instead of modifying the variable tree to remove the multifilter flag, the rendering code now contains logic to decide if the descendant nodes should be displayed. This makes it easier to decide when a node of the variable tree is starable.